real_t eingeführt alle double durch real_t ersetzt um float type einstellen zu können CUDA 64bit atomic Add in cuda-kernel header verschoben

+++ src/OpenCL/kernels/add_mult.cl
@@ -1,4 +1,4 @@
-__kernel void add_mult(__global double* vec1,__global double* vec2, double value, int dim){
+__kernel void add_mult(__global float* vec1,__global float* vec2, float value, int dim){
 	//int id = blockIdx.x * blockDim.x + threadIdx.x;
 	int id = get_global_id(0)
 	if(id < dim){
